Fix a race condition in TestToolService() involving the background protein metadata updater when importing FASTA.
Also refactor usages of StackTraceLogger(), the helper class that was originally used to track down this class of error elsewhere. Cut and pasted copies of code that calls this have been consolidated into a single implementation, ImportFastaDocChangeLogger.
